+++ CHANGELOG.md
@@ -6,11 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/swisscom/powergrr/compare/v0.6.0...master)
 
-<!--
 ### Added
-
+* Add support function `ConvertTo-Hex` for converting an input string into a
+    hex string. This is useful e.g. for hash_entry payloads where the hashes
+    must be converted first to base64 and then to hex.
 ### Changed
-
+* Change encoding in `ConvertFrom-Base64` to unicode.
+<!--
 ### Deprecated
 
 ### Removed